test(app): stop database writers in module tests
