+++ src/app/chat_ordering/tests.rs
@@ -1,10 +1,8 @@
 use super::*;
+use crate::app::test_support::TestApp;
 
-fn app() -> (App<'static>, tempfile::TempDir) {
-    let dir = tempfile::tempdir().unwrap();
-    let app = App::with_data_dir(dir.path(), dir.path());
-    app.db_handler.init();
-    (app, dir)
+fn app() -> TestApp {
+    TestApp::new()
 }
 
 fn message(chat: &wr::JID, id: &str, timestamp: i64) -> wr::Message {
@@ -25,7 +23,7 @@ fn message(chat: &wr::JID, id: &str, timestamp: i64) -> wr::Message {
 
 #[test]
 fn out_of_order_history_is_sorted_oldest_first() {
-    let (mut app, _dir) = app();
+    let mut app = app();
     let chat = wr::JID::from("chat@example.test".to_owned());
     for item in [("newest", 30), ("oldest", 10), ("middle", 20)] {
         app.add_message(message(&chat, item.0, item.1));
@@ -46,7 +44,7 @@ fn out_of_order_history_is_sorted_oldest_first() {
 
 #[test]
 fn equal_timestamps_are_tied_by_message_id() {
-    let (mut app, _dir) = app();
+    let mut app = app();
     let chat = wr::JID::from("chat@example.test".to_owned());
     for id in ["message-c", "message-a", "message-b"] {
         app.add_message(message(&chat, id, 10));
@@ -62,7 +60,7 @@ fn equal_timestamps_are_tied_by_message_id() {
 
 #[test]
 fn new_message_preserves_selected_message_by_id() {
-    let (mut app, _dir) = app();
+    let mut app = app();
     let chat = wr::JID::from("chat@example.test".to_owned());
     app.open_chat = Some(chat.clone());
     for item in [("oldest", 10), ("middle", 20), ("newest", 30)] {
@@ -80,7 +78,7 @@ fn new_message_preserves_selected_message_by_id() {
 
 #[test]
 fn new_message_in_other_chat_leaves_selection_untouched() {
-    let (mut app, _dir) = app();
+    let mut app = app();
     let chat = wr::JID::from("chat@example.test".to_owned());
     let other = wr::JID::from("other@example.test".to_owned());
     app.open_chat = Some(chat.clone());
